requirements.txt
flair
attrs
quickvec